Refactor party implementation
- Move redis functionality out of endpoints module
- Differentiate between a member(ship) and a player
- Add IDs in addition to URLs where appropriate
